Add files via upload
Description:

For this assignment, you will create a RESTful API server using JavaScript and Node.js, as well as a  client to access your own API using Java.  In other words, the server will be written in Node.js and the client will be written in Java.

Requirements:

1) Create a JavaScript file named app.js that will contain your Node.js application.

2) If the client sends an HTTP POST request with a URL of "/logdate",  your server should write the current date to a file named datelog.txt. Each log entry should be appended to existing log entries, and each entry should be separated by a new line.  The server should create a new datelog.txt file if one doesn't exist already.

3) If the client sends an HTTP GET request with a URL of "/dates", your server should return the contents of datelog.txt to the client in an HTTP response.

4) If the client sends any other HTTP requests, the server should reply with a status code of 404.

Server Notes:

You should only require the http and fs modules.  No other npm installs can be required to run your server (although you are welcome to use nodemon).
Include Node.js exception handling as shown in class (i.e. "throw")
Log details of each interaction to the Node.js console.
Client Notes:

Your client should provide an interactive menu that allows the user to either log the current date and on the server, or view the entire contents of the server log.
The client should display a "Resource not available" message if the server returns a status code of 404.